fix: move backups to temporary folder before unpacking/decrypting

+++ .github/scripts/test.sh
@@ -11,13 +11,15 @@ main() {
   docker-compose -f $docker_path/docker-compose.test.yml build --build-arg INSTALL_DOCKER=false
   docker-compose -f $docker_path/docker-compose.test.yml run --rm sut
 
-  # Run tests with Docker installed
-  docker-compose -f $docker_path/docker-compose.test.yml build --build-arg INSTALL_DOCKER=true
-  docker-compose -f $docker_path/docker-compose.test.yml run -v /var/run/docker.sock:/var/run/docker.sock --rm sut
+  if [ "$RUN_ALL" = "true" ]; then
+    # Run tests with Docker installed
+    docker-compose -f $docker_path/docker-compose.test.yml build --build-arg INSTALL_DOCKER=true
+    docker-compose -f $docker_path/docker-compose.test.yml run -v /var/run/docker.sock:/var/run/docker.sock --rm sut
 
-  # Run tests on alpine image
-  docker-compose -f $docker_path/docker-compose.test.yml -f $docker_path/docker-compose.test.alpine.yml build
-  docker-compose -f $docker_path/docker-compose.test.yml -f $docker_path/docker-compose.test.alpine.yml run --rm sut
+    # Run tests on alpine image
+    docker-compose -f $docker_path/docker-compose.test.yml -f $docker_path/docker-compose.test.alpine.yml build
+    docker-compose -f $docker_path/docker-compose.test.yml -f $docker_path/docker-compose.test.alpine.yml run --rm sut
+  fi
 }
 
 SCRIPTDIR=$(dirname "$(readlink -f "$0")")
